pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manifest-guard"
version = "0.2.0"
description = "Production-oriented Kubernetes manifest detector and auto-remediator with remediation memory"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{name = "Your Organization"}]
dependencies = [
  "PyYAML>=6.0",
  "typer>=0.12.0",
  "rich>=13.7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.11.0",
  "build>=1.2.2",
  "pre-commit>=3.7.1"
]

[project.scripts]
manifest-guard = "manifest_guard.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
